Introduced `tkutil:with-hash-table'
The function finds the entry in HASH-TABLE whose key is KEY and
returns the asociated value and T as multiple values.
If there is no such entry, BODY is executed, the result is added to
HASH-TABLE, and the result is returned.

If HASH-TABLE is nil, BODY is always executed and the result is returned.